Fix: calculation of current battery percentage
Fix battery level calculation by using the correct full-charge value.

The full-charge value was being read from the "charge_full_design" file
which holds the nominal capacity of the battery instead of the actual
value when the battery is full. Charge full is the value measured by the
charge gauge circuit after a full discharge and charge and it can be
larger, not all cells are exactly the same, some might have a little
less capacity than nominal, some might have more, and that value will
decrease with use and aging as the cells will hold less charge.